requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.1.1
requests==2.31.0